+++ src/interrupt.rs
@@ -297,6 +297,368 @@ pub fn free<R>(f: impl FnOnce() -> R) -> R {
     f()
 }
 
+#[cfg(feature = "wake-from-isr")]
+pub mod asynch {
+    pub type HalIsrNotification = crate::task::asynch::Notification;
+}
+
+#[cfg(not(feature = "wake-from-isr"))]
+pub mod asynch {
+    use core::{
+        cell::UnsafeCell,
+        ffi::{c_void, CStr},
+        future::Future,
+        num::NonZeroU32,
+        sync::atomic::{AtomicPtr, Ordering},
+        task::{Context, Poll, Waker},
+    };
+
+    use esp_idf_sys::EspError;
+
+    use log::info;
+
+    use crate::{
+        cpu::Core,
+        delay,
+        task::{asynch::Notification, CriticalSection},
+    };
+
+    use super::IsrCriticalSection;
+
+    /// The HAL-global wake runner.
+    /// You should use no more than 64 tasks with it.
+    ///
+    /// `*IsrNotification` instances use this wake runner when they are triggered from an ISR context.
+    pub static HAL_WAKE_RUNNER: WakeRunner<64> = WakeRunner::new(WakeRunnerConfig::new());
+
+    /// Wake runner configuration
+    #[derive(Clone, Debug)]
+    pub struct WakeRunnerConfig {
+        pub task_name: &'static CStr,
+        pub task_stack_size: usize,
+        pub task_priority: u8,
+        pub task_pin_to_core: Option<Core>,
+    }
+
+    impl WakeRunnerConfig {
+        pub const fn new() -> Self {
+            Self {
+                task_name: unsafe { CStr::from_bytes_with_nul_unchecked(b"WakeRunner\0") },
+                task_stack_size: 3084,
+                task_priority: 11,
+                task_pin_to_core: None,
+            }
+        }
+    }
+
+    impl Default for WakeRunnerConfig {
+        fn default() -> Self {
+            Self::new()
+        }
+    }
+
+    /// WakeRunner is a utility allowing `Waker` instances to be awoken fron an ISR context.
+    ///
+    /// General problem:
+    /// In an interrupt, using Waker instances coming from generic executors is impossible,
+    /// because these are not designed with an ISR-safety in mind.
+    ///
+    /// Waking a waker means that its task would be scheduled on the executor queue, which might involve
+    /// allocation, and/or synchronization primitives which are not safe to use from an ISR context.
+    ///
+    /// Similarly, dropping a waker might also drop the executor task, resulting in a deallocation, which is also
+    /// not safe in an ISR context.
+    ///
+    /// These problems are alleviated by replacing direct `waker.wake()` calls to `WakerRunner::wake(waker)`.
+    /// What `WakeRunner::wake` does is to push the waker into a bounded queue and then notify a hidden FreeRTOS task.
+    /// Once the FreeRTOS task gets awoken, it wakes all wakers scheduled on the bounded queue and empties the queue.
+    pub struct WakeRunner<const N: usize> {
+        wakers_cs: IsrCriticalSection,
+        wakers: UnsafeCell<heapless::Deque<Waker, N>>,
+        task_cs: CriticalSection,
+        task: AtomicPtr<crate::sys::tskTaskControlBlock>,
+        task_config: WakeRunnerConfig,
+    }
+
+    impl<const N: usize> WakeRunner<N> {
+        /// Create a new `WakeRunner` instance.
+        pub const fn new(config: WakeRunnerConfig) -> Self {
+            Self {
+                wakers_cs: IsrCriticalSection::new(),
+                wakers: UnsafeCell::new(heapless::Deque::new()),
+                task_cs: CriticalSection::new(),
+                task: AtomicPtr::new(core::ptr::null_mut()),
+                task_config: config,
+            }
+        }
+
+        /// Returns `true` if the wake runner is started.
+        pub fn is_started(&self) -> bool {
+            !self.task.load(Ordering::SeqCst).is_null()
+        }
+
+        /// Starts the wake runner. Returns `false` if it had been already started.
+        pub fn start(&'static self) -> Result<bool, EspError> {
+            let _guard = self.task_cs.enter();
+
+            if self.task.load(Ordering::SeqCst).is_null() {
+                let task = unsafe {
+                    crate::task::create(
+                        Self::task_run,
+                        self.task_config.task_name,
+                        self.task_config.task_stack_size,
+                        self as *const _ as *const c_void as *mut _,
+                        self.task_config.task_priority,
+                        self.task_config.task_pin_to_core,
+                    )?
+                };
+
+                self.task.store(task as _, Ordering::SeqCst);
+
+                info!("WakeRunner {:?} started.", self.task_config.task_name);
+
+                Ok(true)
+            } else {
+                Ok(false)
+            }
+        }
+
+        /// Stops the wake runner. Returns `false` if it had been already stopped.
+        pub fn stop(&self) -> bool {
+            let _guard = self.task_cs.enter();
+
+            let task = self.task.swap(core::ptr::null_mut(), Ordering::SeqCst);
+
+            if !task.is_null() {
+                unsafe {
+                    crate::task::destroy(task as _);
+                }
+
+                info!("WakeRunner {:?} stopped.", self.task_config.task_name);
+
+                true
+            } else {
+                false
+            }
+        }
+
+        /// Schedules a waker to be awoken by the hidden FreeRTOS task running in the background.
+        /// If not called from within an ISR context, calls `waker.wake()` directly instead of scheduling the waker.
+        /// NOTE: If the wake runner is not started yet, scheduling fron an ISR content would fail silently.
+        ///
+        /// This and only this method is safe to call from an ISR context.
+        pub fn schedule(&self, waker: Waker) {
+            if super::active() {
+                self.wakers(|wakers| {
+                    let earlier_waker = wakers.iter_mut().find(|a_waker| a_waker.will_wake(&waker));
+
+                    if let Some(earlier_waker) = earlier_waker {
+                        *earlier_waker = waker;
+                    } else if wakers.push_back(waker).is_err() {
+                        panic!("WakeRunner queue overflow");
+                    }
+
+                    let task = self.task.load(Ordering::SeqCst);
+
+                    if !task.is_null() {
+                        unsafe {
+                            crate::task::notify_and_yield(task as _, NonZeroU32::new(1).unwrap());
+                        }
+                    }
+                })
+            } else {
+                waker.wake();
+            }
+        }
+
+        fn run(&self) {
+            loop {
+                loop {
+                    let waker = self.wakers(|wakers| wakers.pop_front());
+
+                    if let Some(waker) = waker {
+                        waker.wake();
+                    } else {
+                        break;
+                    }
+                }
+
+                crate::task::wait_notification(delay::BLOCK);
+            }
+        }
+
+        fn wakers<F: FnOnce(&mut heapless::Deque<Waker, N>) -> R, R>(&self, f: F) -> R {
+            // if super::active() {
+            //     let wakers = unsafe { self.wakers.get().as_mut().unwrap() };
+
+            //     f(wakers)
+            // } else {
+            let _guard = self.wakers_cs.enter();
+
+            let wakers = unsafe { self.wakers.get().as_mut().unwrap() };
+
+            f(wakers)
+            // }
+        }
+
+        extern "C" fn task_run(ctx: *mut c_void) {
+            let this =
+                unsafe { (ctx as *mut WakeRunner<N> as *const WakeRunner<N>).as_ref() }.unwrap();
+
+            this.run();
+        }
+    }
+
+    impl<const N: usize> Drop for WakeRunner<N> {
+        fn drop(&mut self) {
+            self.stop();
+        }
+    }
+
+    unsafe impl<const N: usize> Send for WakeRunner<N> {}
+    unsafe impl<const N: usize> Sync for WakeRunner<N> {}
+
+    /// Single-slot lock-free signaling primitive supporting signalling with a `u32` bit-set.
+    /// A variation of the `Notification` HAL primitive which is however safe to be notified from an ISR context.
+    ///
+    /// It is useful for sending data between an ISR routine (or a regular task context) and an async task when the
+    /// receiver only cares about the latest data, and therefore it's fine to "lose" messages.
+    /// This is often the case for "state" updates.
+    ///
+    /// The sending part of the primitive is non-blocking and ISR-safe, so it can be called from anywhere.
+    ///
+    /// Similar in spirit to the ESP-IDF FreeRTOS task notifications in that it is light-weight and operates on bit-sets,
+    /// but for synchronization between an asynchronous task, and another one, which might be blocking or asynchronous.
+    pub struct IsrNotification<const N: usize> {
+        inner: Notification,
+        runner: &'static WakeRunner<N>,
+    }
+
+    impl<const N: usize> IsrNotification<N> {
+        /// Creates a new `IsrNotification`.
+        /// This method is safe to call from an ISR context, yet such use cases should not normally occur in practice.
+        pub const fn new(runner: &'static WakeRunner<N>) -> Self {
+            Self {
+                inner: Notification::new(),
+                runner,
+            }
+        }
+
+        /// Marks the least significant bit (bit 0) in this `IsrNotification` as nofified.
+        /// This method is safe to call from an ISR context.
+        /// Returns `true` if there was a registered waker which got awoken.
+        pub fn notify_lsb(&self) -> bool {
+            self.notify(NonZeroU32::new(1).unwrap())
+        }
+
+        /// Marks the supplied bits in this `IsrNotification` as notified.
+        /// This method is safe to call from an ISR context.
+        /// Returns `true` if there was a registered waker which got awoken.
+        pub fn notify(&self, bits: NonZeroU32) -> bool {
+            if let Some(waker) = self.inner.notify_waker(bits) {
+                self.runner.schedule(waker);
+
+                true
+            } else {
+                false
+            }
+        }
+
+        /// Clears the state of this notification by removing any registered waker and setting all bits to 0.
+        /// This method is NOT safe to call from an ISR context.
+        pub fn reset(&self) {
+            self.inner.reset();
+        }
+
+        /// Future that completes when this `IsrNotification` has been notified.
+        /// This method is NOT safe to call from an ISR context.
+        #[allow(unused)]
+        pub fn wait(&self) -> impl Future<Output = NonZeroU32> + '_ {
+            self.runner.start().unwrap();
+
+            self.inner.wait()
+        }
+
+        /// Non-blocking method to check whether this notification has been notified.
+        /// This method is NOT safe to call from an ISR context.
+        pub fn poll_wait(&self, cx: &Context<'_>) -> Poll<NonZeroU32> {
+            self.runner.start().unwrap();
+
+            self.inner.poll_wait(cx)
+        }
+    }
+
+    /// Single-slot lock-free signaling primitive supporting signalling with a `u32` bit-set.
+    /// A variation of the `IsrNotification` HAL primitive which is however safe to be notified from an ISR context.
+    /// The difference between this primitive and `IsrNotification` is that this one is hard-wired to the
+    /// global HAL wake runner (`HAL_WAKE_RUNNER`) and is thus occupying less space.
+    ///
+    /// It is useful for sending data between an ISR routine (or a regular task context) and an async task when the
+    /// receiver only cares about the latest data, and therefore it's fine to "lose" messages.
+    /// This is often the case for "state" updates.
+    ///
+    /// The sending part of the primitive is non-blocking and ISR-safe, so it can be called from anywhere.
+    ///
+    /// Similar in spirit to the ESP-IDF FreeRTOS task notifications in that it is light-weight and operates on bit-sets,
+    /// but for synchronization between an asynchronous task, and another one, which might be blocking or asynchronous.
+    pub struct HalIsrNotification {
+        inner: Notification,
+    }
+
+    impl HalIsrNotification {
+        /// Creates a new `HalIsrNotification`.
+        /// This method is safe to call from an ISR context, yet such use cases should not normally occur in practice.
+        pub const fn new() -> Self {
+            Self {
+                inner: Notification::new(),
+            }
+        }
+
+        /// Marks the least significant bit (bit 0) in this `IsrNotification` as nofified.
+        /// This method is safe to call from an ISR context.
+        /// Returns `true` if there was a registered waker which got awoken.
+        pub fn notify_lsb(&self) -> bool {
+            self.notify(NonZeroU32::new(1).unwrap())
+        }
+
+        /// Marks the supplied bits in this `HalIsrNotification` as notified.
+        /// This method is safe to call from an ISR context.
+        /// Returns `true` if there was a registered waker which got awoken.
+        pub fn notify(&self, bits: NonZeroU32) -> bool {
+            if let Some(waker) = self.inner.notify_waker(bits) {
+                HAL_WAKE_RUNNER.schedule(waker);
+
+                true
+            } else {
+                false
+            }
+        }
+
+        /// Clears the state of this notification by removing any registered waker and setting all bits to 0.
+        /// This method is NOT safe to call from an ISR context.
+        pub fn reset(&self) {
+            self.inner.reset();
+        }
+
+        /// Future that completes when this `HalIsrNotification` has been notified.
+        /// This method is NOT safe to call from an ISR context.
+        #[allow(unused)]
+        pub fn wait(&self) -> impl Future<Output = NonZeroU32> + '_ {
+            HAL_WAKE_RUNNER.start().unwrap();
+
+            self.inner.wait()
+        }
+
+        /// Non-blocking method to check whether this notification has been notified.
+        /// This method is NOT safe to call from an ISR context.
+        pub fn poll_wait(&self, cx: &Context<'_>) -> Poll<NonZeroU32> {
+            HAL_WAKE_RUNNER.start().unwrap();
+
+            self.inner.poll_wait(cx)
+        }
+    }
+}
+
 #[cfg(feature = "embassy-sync")]
 pub mod embassy_sync {
     use core::marker::PhantomData;