+++ src/lib.rs
@@ -71,8 +71,6 @@ pub mod peripheral;
 pub mod peripherals;
 pub mod prelude;
 #[cfg(not(feature = "riscv-ulp-hal"))]
-mod private;
-#[cfg(not(feature = "riscv-ulp-hal"))]
 pub mod reset;
 #[cfg(not(feature = "riscv-ulp-hal"))]
 pub mod rmt;